backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
pyotp==2.9.0
qrcode[pil]==7.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
requests==2.31.0
aiofiles==23.2.1
pyclamd==0.4.0
